sequential id to uuid

+++ src/main/java/io/userauth/models/Role.java
@@ -13,8 +13,6 @@ import lombok.Setter;
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private int id;
     private String name;
-
 }
